Split out http and aws libraries to prepare for separate libs Return nil for invalid span so that otel defaults does not map
