+++ pkg/trace/interpreter/interpreters/open-telemetry/open_telemetry_span_builder.go
@@ -1,4 +1,4 @@
-package aws
+package opentelemetry
 
 import (
 	"fmt"
@@ -6,14 +6,15 @@ import (
 )
 
 // OpenTelemetrySpanBuilder An generic function to map Open Telemetry AWS service traces to a format that STS understands
-func OpenTelemetrySpanBuilder(span *pb.Span, spanKind string, spanEvent string, serviceID string, serviceType string, serviceLayer string, serviceDomain string, urn string, arn string) {
+func OpenTelemetrySpanBuilder(span *pb.Span, spanKind string, spanEvent string, serviceID string, serviceType string, serviceLayer string, serviceDomain string, urn string, identifier string) {
 	var mappingKey = "open.telemetry"
 
 	span.Type = serviceType
 	span.Resource = fmt.Sprintf("aws.%s", serviceID)
 	span.Service = fmt.Sprintf("%s.%s", mappingKey, serviceID)
 
-	span.Meta["sts.service.identifiers"] = arn
+	span.Meta["sts.service.identifiers"] = identifier
+	span.Meta["span.serviceURN"] = urn
 
 	span.Meta["service"] = span.Service
 	span.Meta["layer"] = serviceLayer
@@ -22,7 +23,6 @@ func OpenTelemetrySpanBuilder(span *pb.Span, spanKind string, spanEvent string, 
 	span.Meta["span.serviceType"] = serviceType
 	span.Meta["span.serviceName"] = fmt.Sprintf("%s.%s.%s", mappingKey, serviceID, spanEvent)
 	span.Meta["span.kind"] = spanKind
-	span.Meta["span.serviceURN"] = urn
 	span.Meta["span.layer"] = serviceLayer
 	span.Meta["span.domain"] = serviceDomain
 }
